// src/app/api/swagger/models/api-market-entity-type-enum.ts
/* tslint:disable */
/* eslint-disable */
export enum ApiMarketEntityTypeEnum {
  Offer = 0,
  Trade = 1,
  Deal = 2,
  Contract = 3,
  Application = 4,
  Organization = 5,
  Discussion = 6,
  TradeLot = 7,
  TradePlan = 8,
  Feedback = 9,
  RequirementRequest = 10,
  ApprovalRequest = 11,
  OrderAgreementHistory = 12,
  FinancialTransaction = 13,
  Characteristic = 14,
  PriceList = 15,
  PriceListImportTask = 16,
  Cart = 17,
  Tenant = 18,
  LotItem = 19
}